feat: :sparkles: /leaderboard 각 탭 SegmentedControl + 더미
