--- app/src/pages/LeaderBoardPage/tabs/CorrectionPointRankTab.tsx
@@ -1,3 +1,0 @@
-export const CorrectionPointRankTab = () => {
-  return <>CorrectionPointRankTab</>;
-};
