--- app/src/pages/LeaderBoardPage/tmp/LevelRank.tsx
@@ -1,54 +1,0 @@
-import { gql } from '@/__generated__';
-import { Loader } from '@/components/common';
-import {
-  ApolloBadRequest,
-  ApolloNotFound,
-} from '@/components/elements/DashboardContentView';
-import { Rank } from '@/components/elements/DashboardContentView/Rank';
-import { BelowTablet, Desktop } from '@/utils/responsive/Device';
-import type { RankItemType } from '@/utils/types/Rank';
-import { useQuery } from '@apollo/client';
-
-const GET_LEVEL_RANK = gql(/* GraphQL */ `
-  query GetLevelRank {
-    getHomePage {
-      levelRank {
-        userPreview {
-          id
-          login
-          imgUrl
-        }
-        value
-      }
-    }
-  }
-`);
-
-export const LevelRank = () => {
-  const { loading, error, data } = useQuery(GET_LEVEL_RANK);
-
-  if (loading) return <Loader />;
-  if (error) return <ApolloBadRequest msg={error.message} />;
-  if (!data) return <ApolloNotFound />;
-
-  const { levelRank } = data.getHomePage;
-  const unit = '';
-
-  const rankList: RankItemType[] = levelRank.map(({ userPreview, value }) => ({
-    id: userPreview.id,
-    name: userPreview.login,
-    value: value,
-    imgUrl: userPreview.imgUrl,
-  }));
-
-  return (
-    <>
-      <Desktop>
-        <Rank rankList={rankList} cnt={3} unit={unit} />
-      </Desktop>
-      <BelowTablet>
-        <Rank rankList={rankList} showImg={false} cnt={3} unit={unit} />
-      </BelowTablet>
-    </>
-  );
-};
